+++ src/main/java/com/hmdandelion/project_1410002/sales/domain/repository/client/ClientRepo.java
@@ -1,7 +1,7 @@
 package com.hmdandelion.project_1410002.sales.domain.repository.client;
 
 import com.hmdandelion.project_1410002.sales.domain.entity.client.Client;
-import com.hmdandelion.project_1410002.sales.model.ClientStatus;
+import com.hmdandelion.project_1410002.sales.domain.type.ClientStatus;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
