+++ src/main/java/com/hmdandelion/project_1410002/sales/domain/repository/estimate/EstimateProductRepo.java
@@ -1,4 +1,4 @@
-package com.hmdandelion.project_1410002.sales.domain.repository;
+package com.hmdandelion.project_1410002.sales.domain.repository.estimate;
 
 import com.hmdandelion.project_1410002.sales.domain.entity.estimate.EstimateProduct;
 import org.springframework.data.jpa.repository.JpaRepository;
