+++ src/main/java/com/hmdandelion/project_1410002/sales/domain/type/ClientStatus.java
@@ -1,0 +1,5 @@
+package com.hmdandelion.project_1410002.sales.domain.type;
+
+public enum ClientStatus {
+    ACTIVE, DELETED
+}
