+++ src/main/java/com/hmdandelion/project_1410002/sales/domain/type/EstimateStatus.java
@@ -1,0 +1,5 @@
+package com.hmdandelion.project_1410002.sales.domain.type;
+
+public enum EstimateStatus {
+    ACTIVE, DELETED
+}
