+++ src/main/java/com/hmdandelion/project_1410002/sales/dto/response/EstimatesResponse.java
@@ -1,8 +1,7 @@
 package com.hmdandelion.project_1410002.sales.dto.response;
 
 import com.fasterxml.jackson.annotation.JsonFormat;
-import com.hmdandelion.project_1410002.sales.domain.entity.estimate.Estimate;
-import com.hmdandelion.project_1410002.sales.model.EstimateStatus;
+import com.hmdandelion.project_1410002.sales.domain.type.EstimateStatus;
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
