--- src/main/java/com/hmdandelion/project_1410002/sales/model/ClientStatus.java
@@ -1,5 +1,0 @@
-package com.hmdandelion.project_1410002.sales.model;
-
-public enum ClientStatus {
-    ACTIVE, DELETED
-}
