--- src/main/java/com/hmdandelion/project_1410002/sales/model/ClientType.java
@@ -1,5 +1,0 @@
-package com.hmdandelion.project_1410002.sales.model;
-
-public enum ClientType {
-    RAW_MATERIALS, PRODUCTS
-}
