--- src/main/java/com/hmdandelion/project_1410002/sales/model/EstimateStatus.java
@@ -1,5 +1,0 @@
-package com.hmdandelion.project_1410002.sales.model;
-
-public enum EstimateStatus {
-    ACTIVE, DELETED
-}
