+++ src/main/java/com/hmdandelion/project_1410002/sales/service/EstimateService.java
@@ -2,18 +2,16 @@ package com.hmdandelion.project_1410002.sales.service;
 
 import com.hmdandelion.project_1410002.common.exception.NotFoundException;
 import com.hmdandelion.project_1410002.common.exception.type.ExceptionCode;
-import com.hmdandelion.project_1410002.sales.domain.entity.client.Client;
 import com.hmdandelion.project_1410002.sales.domain.entity.estimate.Estimate;
 import com.hmdandelion.project_1410002.sales.domain.entity.estimate.EstimateProduct;
-import com.hmdandelion.project_1410002.sales.domain.repository.EstimateProductRepo;
 import com.hmdandelion.project_1410002.sales.domain.repository.client.ClientRepo;
 import com.hmdandelion.project_1410002.sales.domain.repository.estimate.EstimateRepo;
 import com.hmdandelion.project_1410002.sales.dto.request.EstimateCreateRequest;
 import com.hmdandelion.project_1410002.sales.dto.request.EstimateUpdateRequest;
 import com.hmdandelion.project_1410002.sales.dto.response.EstimateResponse;
 import com.hmdandelion.project_1410002.sales.dto.response.EstimatesResponse;
-import com.hmdandelion.project_1410002.sales.model.ClientStatus;
-import com.hmdandelion.project_1410002.sales.model.ClientType;
+import com.hmdandelion.project_1410002.sales.domain.type.ClientStatus;
+import com.hmdandelion.project_1410002.sales.domain.type.ClientType;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
